Ensure coordinates are passed to MapBox as lng , lat
MapBox cites GeoJSON's position format dictating longitude, latitude:

https://docs.mapbox.com/mapbox-gl-js/api/geography/#lnglatlike
https://geojson.org/geojson-spec.html#positions

ISO 6709 dictates latitude, longitude:

https://en.wikipedia.org/wiki/ISO_6709#Items

For now, we're choosing to store coordinate as lat, lng, so we need to reverse them
when passing coordinates to Mapbox.

This wasn't seen on the initial Explore Homer prototype, because the data from
Plieades was also returned as longitude, latitude.

Our reason for choosing lat, lng is conformance to the TEI spec:

https://tei-c.org/release/doc/tei-p5-doc/en/html/ref-geo.html